refactor: section 패키지 생성
