+++ src/main/java/wooteco/subway/domain/section/SectionAddStatus.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import java.util.Arrays;
 import java.util.List;
