+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -8,9 +8,9 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.section.Sections;
 import wooteco.subway.dto.request.SectionRequest;
 import wooteco.subway.exception.LineNotFoundException;
 import wooteco.subway.exception.StationNotFoundException;
